+++ packages/cli/uni-builder/package.json
@@ -71,6 +71,7 @@
     "lodash": "^4.17.21",
     "magic-string": "0.30.12",
     "oxc-parser": "^0.36.0",
+    "@swc/core": "1.3.42",
     "picocolors": "^1.0.1",
     "postcss": "^8.4.35",
     "postcss-custom-properties": "13.3.12",
@@ -89,25 +90,16 @@
     "webpack-subresource-integrity": "5.1.0"
   },
   "devDependencies": {
-<<<<<<< HEAD
-    "@rsbuild/plugin-webpack-swc": "1.0.9",
-=======
     "@rsbuild/plugin-webpack-swc": "~1.0.3",
     "@modern-js/types": "workspace:*",
->>>>>>> a5b30dbfec (feat: support rsc)
     "@scripts/build": "workspace:*",
     "@scripts/vitest-config": "workspace:*",
     "@types/es-module-lexer": "^0.4.3",
     "@types/html-minifier-terser": "^7.0.2",
     "@types/lodash": "^4.14.202",
-<<<<<<< HEAD
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-=======
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-server-dom-webpack": "19.0.0",
->>>>>>> a5b30dbfec (feat: support rsc)
     "terser": "^5.31.1",
     "typescript": "^5.3.0"
   },
